installer descriptor editor: add key module
